moved stuff around, using HSQLDB for more tests instead of MySQL or Postgres (where not needed). MassIndexer in genericjpa should be deprecated and replaced by a generic version of the original one
