+++ genericjpa/jpa/src/test/resources/META-INF/persistence.xml
@@ -45,7 +45,7 @@
         </properties>
     </persistence-unit>
 
-    <persistence-unit name="Hibernate_MySQL"
+    <persistence-unit name="Hibernate_HSQLDB"
                       transaction-type="RESOURCE_LOCAL">
         <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
         <class>org.hibernate.search.genericjpa.test.jpa.entities.Sorcerer</class>
@@ -55,11 +55,12 @@
 
         <class>org.hibernate.search.genericjpa.test.jpa.entities.MultipleColumnsIdEntity</class>
         <properties>
-            <property name="javax.persistence.jdbc.driver" value="com.mysql.jdbc.Driver" />
-            <property name="javax.persistence.jdbc.url" value="jdbc:mysql://localhost:3306/testingdb" />
-            <property name="javax.persistence.jdbc.user" value="hibernate_user" />
-            <property name="javax.persistence.jdbc.password" value="hibernate_password" />
-            <property name="hibernate.dialect" value="org.hibernate.dialect.MySQL5Dialect" />
+
+            <property name="javax.persistence.jdbc.driver" value="org.hsqldb.jdbcDriver"/>
+            <property name="javax.persistence.jdbc.url" value="jdbc:hsqldb:mem:test"/>
+            <property name="javax.persistence.jdbc.user" value="hibernate_user"/>
+            <property name="javax.persistence.jdbc.password" value="hibernate_password"/>
+            <property name="hibernate.dialect" value="org.hibernate.dialect.HSQLDialect" />
             <property name="hibernate.hbm2ddl.auto" value="create-drop" />
         </properties>
     </persistence-unit>
